Correct special net NPt dst prefix handling. Fixes #13240
Disables prefix length drop-down when using a special net (e.g. track6
delegated prefix) because that already has its own prefix length.

Being able to specify a custom prefix was of dubious use and served to
be more confusing to users, so disable that and just take the prefix
from the delegation.