Work around for empty config tags. Fixes #13564.
Related to issue #13642, the serialization of route maps was broken for a time
in such a way that the frrglobalroutemaps tag would contain only a single empty
config tag. The xml parser element handler considers 'config' to be a 'list
tag', meaning that multiple of them are to be expected. The way it handles list
tags is to create an array of the tag name (if it doesn't exist), and tack the
index of the end of the array into the path. If the tag is closed without deeper
tags being processed, an empty string is stored in the array element. For config
tags, this is never what we want as throughout the codebase elements in a config
array are expected to be arrays.

To work around the condition caused by the incorrect population of sorted route
maps, this commit introduces a test to the end element handler that pops the
last element off a config array if the last element is not an array itself. This
will result in empty config tags simply being skipped as they are to be
considered malformed.