Prevent get_pf_rules() from indexing a string error. Fixes #13660
pfSense_get_pf_rules(), which populates the $rulescnt variable sent to
get_pf_rules(), will terminate its loop fetching rules if a call to
pfctl_get_rule() returns nonzero, adding to the associative array return value
and element keyed 'error' that is a scalar. The original loop over the return
value in get_pf_rules() asserts that each element in the result array is an
array of rule attributes, and if this error value exists it will trigger an
exception attempting to index a string. This commit restores the behavior prior
to the upgrade to php 8.1 by skipping the error value. The rule list returned by
get_pf_rules() may be incomplete if the error value exists.