Correct NAT state regex matching for rrd graphs. Fixes #13966
The original regex was ambiguous for posix REs, and some basic RE application
may have been done for *? which is not valid posix extended RE syntax. The match
also appears to use an old format for distinguishing NAT rules that is no longer
current. This change alters the REs to match NAT rules according to the current
pfctl output.