Safely update efi loader from pfSense-boot. Fixes #14045
Alter pfSense-boot +INSTALL script to safely update the efi loader from
/boot/loader.efi. The update is now done with 'install -S' to ensure that any IO
failure does not leave an incomplete loader in the efi partition.

Additionally, the mounted ${PKG_ROOTDIR}/boot/efi is ensured to be unmounted by
an EXIT signal trap when it is successfully mounted, and the expected directory
structure is created in the efi filesystem if it does not exist.