Improve alias sorting (again). Issue #14015
asort does not handle natural case-insensitive sorting of
multi-dimensional arrays properly, so it needs a custom sort but also
has to preserve the index keys or else making changes in the GUI won't
use the correct IDs.

The recent rector pass also removed sort-on-save which we didn't really
need, but it was nice to have the aliases sorted in the ruleset so we
can preserve that using the same function as well.