Remove failover states using only the gateway label
Interfaces were previously specified since the inbound state needs to
be killed (due to route-to) for the connection to restart on the
preferred gateway. This interrupts connections already established on
the preferred gateway. To avoid that interruption, the interface can be
omitted by adding the current gateway tier to the gateway label.
States retain the labels from rules that created them even after the
ruleset changes.