Updated much faster log.widget Bug #12673
This widget is to replace the actual log.widget.. Loading the log in one go does speed up the widget about 20 times. I left debug statements in the code (for now) so that it is easy to folow what is happening. For more detail see Bug #12673: Firewall Logs Widget fails to update at intervals below 5 seconds. - pfSense - pfSense bugtracker