Remove potential direct config references when displaying form rows
The first eval() change removes the reference and has no functional
effect given that $pkg_source_txt is not modified. While here, catch
any exceptions as well; before PHP 7, eval() would return false on
parse errors and execution continued normally.

Regarding the second eval() change:
The inclusion of the reference results in $config containing the
default values as determined by 'show_disable_value'. The default
values are displayed as an item for selection on the form list. Since
no configuration write is performed until the form is submitted, the
inclusion of default values in $config is discarded (by omission)
unless included in POST. Hence the reference may aslo be removed.