Support IPv6 requests for any DDNS service. Implement #11177
Introduce get_dyndns_source_address() to find the source address to be
used when sending a DDNS request. It supports both IPv4 and IPv6 -
determined by the configured gateway and addresses on the interface.

Update dyndnsCheckIP() to use get_dyndns_source_address() for IPv6
support. Also update the IP pattern to improve matching potential.

Update the DDNS class to use the new code for IPv6 requests.