Reduce rc.newwanipv6 actions on RENEW
This change passes the dhcp6c REASON to the script. The script uses this
to retain the previous behavior for REASONs other than RENEW. If the
reason is RENEW, action is only taken when there has been a change. Also
include the REASON in the info log message.

Followup to 646389402feb2dd94171d7c81d4be67feef4f8d8.