Fix double rc.newwanipv6 execution on if_pppoe
When using if_pppoe and DHCPv6 over the PPP link, both the devd handler script
and dhcp6c execute /etc/rc.newwanipv6 as a result of an address assignment,
which is not fatal but does cause some extra delay as the filter is reloaded and
packages are restarted more than necessary. This is fixed by limiting the
pppoe-handler script to responding to a new IPv6 address only when the logical
interface is not configured for DHCPv6, allowing dhcp6c to manage lease changes
itself. All other IPv6 configurations will generate a ADDR_ADD devd event on
address assignment that is handled by the pppoe-handler script.