Always bring if_pppoe interfaces down before destroying them
For kernel if_pppoe interfaces, when the interface is destroyed it is removed
immediately without sending a Term-Request. This leaves the other end of the PPP
tunnel established until it times out from lack of response. In the interim, if
the if_pppoe interface attempts to connect again it will fail if the service
side does not allow multiple concurrent logins. This is remedied by downing the
interface prior to destruction.