Fix rule labels for user rules in firewall log GUI
Previously the firewall log page checked for "USER_RULE: " in rule labels
to add a user icon to appropriate rules. Now that all rule descriptions
use the same label prefix (i.e. key), a new way is needed to determine if
a rule is considered a user rule. Introduce the "user_rule" label tag for
this purpose.