Use the IPv6 interface address as the OpenVPN server gateway. Fix #16351
In OpenVPN server configurations OpenVPN sets the environment variable
"ifconfig_ipv6_remote" when an IPv6 tunnel network is specified. This is
not the case for IPv4 and its respective variables. These variables are
later used to create a gateway for assigned interfaces in which case the
gateway address should always be the local interface address.