RFC2136 DDNS: use get_request_source_address()
When the RFC2136 config uses the option "usepublicip" the source address
is found by calling the function dyndnsCheckIP() which in turn calls
get_request_source_address(). A different code path is followed when
"usepublicip" is not used which can result in the wrong source address
being returned (e.g. when the interface is a failover gateway group). This
change aligns the logic used for determining the request address
regardless of the "usepublicip" option (except for "_stf" interfaces).