Allow using interface subnet macros which only have VIPs. Implement #16613
Remove the check for a primary address when generating a pf rule's
interface subnet macro. And allow the rule to be generated if the macro
contains any address or VIP. This behavior aligns with the table output
shown in the WebGUI when hovering over the alias used in a rule.